+++ lib/providers/firebase_cache_provider.g.dart
@@ -121,4 +121,4 @@ final class FirebaseCacheProvider
   }
 }
 
-String _$firebaseCacheHash() => r'5f67df1a526aff39b5a122585b46608e6609b098';
+String _$firebaseCacheHash() => r'527db973ec048587e07e0c21ac04763102bee1b0';